빌드 및 실행 성공. PCH 추가

+++ Engine_SOURCE/Input.cpp
@@ -1,3 +1,5 @@
+#include "EnginePCH.h"
+
 #include "Input.h"
 #include "Application.h"
 
@@ -78,8 +80,8 @@ namespace mh
 			POINT mousePos = {};
 			GetCursorPos(&mousePos);
 			ScreenToClient(application.GetHwnd(), &mousePos);
-			mMousPosition.x = mousePos.x;
-			mMousPosition.y = mousePos.y;
+			mMousPosition.x = static_cast<float>(mousePos.x);
+			mMousPosition.y = static_cast<float>(mousePos.y);
 		}
 		else
 		{
